Revert "Script Flink job submission"
This reverts commit 95ccab96ab492604a536213813153a7b246e3296.

It would be cool to submit jobs with a script but at this point we
cannot figure out how to make it behave in the same way as the web UI
submission. The main difference is that jobs submitted via the web UI
can be rerun through a REST API. This is crucial for us at the moment.